Move conversion into sub-package
