+++ open_web_calendar/calendars/__init__.py
@@ -3,12 +3,14 @@
 # SPDX-License-Identifier: GPL-2.0-only
 """A unified interface to calendars from different sources."""
 
+from .base import Calendars
 from .caldav import CalDAVCalendars
 from .errors import InvalidCalendars
 from .ics import ICSCalendars
 
 __all__ = [
     "CalDAVCalendars",
+    "Calendars",
     "ICSCalendars",
     "InvalidCalendars",
 ]
